使用sinosdx-cloud 1.0.2版本

+++ sinosdx-service/service-api-management/src/main/java/com/sinosdx/service/management/controller/dto/ApplicationNumDTO.java
@@ -8,6 +8,6 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class ApplicationNumDTO {
     private Integer subscribedNum;
-    private Integer apiNum;
+    private Long apiNum;
     private Integer applicationNum;
 }
